+++ Planning/Creneau/Controller.php
@@ -1,18 +1,18 @@
 <?php
-namespace LibertAPI\Components\Planning;
+namespace LibertAPI\Planning\Creneau;
 
 use LibertAPI\Tools\Exceptions\MissingArgumentException;
 use Psr\Http\Message\ServerRequestInterface as IRequest;
 use Psr\Http\Message\ResponseInterface as IResponse;
 
 /**
- * Contrôleur de plannings
+ * Contrôleur des creneaux de plannings
  *
  * @author Prytoegrian <prytoegrian@protonmail.com>
  * @author Wouldsmina
  *
  * @since 0.1
- * @see \Tests\Units\Components\Planning\Controller
+ * @see \Tests\Units\Planning\Controller
  *
  * Ne devrait être contacté que par le routeur
  * Ne devrait contacter que le Planning\Repository
@@ -28,18 +28,16 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
      *
      * @param IRequest $request Requête Http
      * @param IResponse $response Réponse Http
-     * @param array $arguments Arguments de route
      *
      * @return IResponse
-     * @throws \Exception en cas d'erreur inconnue (fallback, ne doit pas arriver)
      */
     public function get(IRequest $request, IResponse $response, array $arguments)
     {
-        if (!isset($arguments['planningId'])) {
-            return $this->getList($request, $response);
+        if (!isset($arguments['creneauId'])) {
+            return $this->getList($request, $response, (int) $arguments['planningId']);
         }
 
-        return $this->getOne($response, (int) $arguments['planningId']);
+        return $this->getOne($response, (int) $arguments['creneauId'], (int) $arguments['planningId']);
     }
 
     /**
@@ -47,25 +45,36 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
      *
      * @param IResponse $response Réponse Http
      * @param int $id ID de l'élément
+     * @param int $planningId Contrainte de recherche sur le planning
      *
      * @return IResponse, 404 si l'élément n'est pas trouvé, 200 sinon
      * @throws \Exception en cas d'erreur inconnue (fallback, ne doit pas arriver)
      */
-    private function getOne(IResponse $response, $id)
+    private function getOne(IResponse $response, $id, $planningId)
     {
+        $code = -1;
+        $data = [];
         try {
-            $planning = $this->repository->getOne($id);
+            $creneau = $this->repository->getOne($id, $planningId);
             $code = 200;
             $data = [
                 'code' => $code,
                 'status' => 'success',
                 'message' => '',
-                'data' => $this->buildData($planning),
+                'data' => $this->buildData($creneau),
             ];
 
             return $response->withJson($data, $code);
         } catch (\DomainException $e) {
-            return $this->getResponseNotFound($response, 'Element « plannings#' . $id . ' » is not a valid resource');
+            $code = 404;
+            $data = [
+                'code' => $code,
+                'status' => 'error',
+                'message' => 'Not Found',
+                'data' => 'Element « creneaux#' . $id . ' » is not a valid resource',
+            ];
+
+            return $response->withJson($data, $code);
         } catch (\Exception $e) {
             throw $e;
         }
@@ -76,19 +85,20 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
      *
      * @param IRequest $request Requête Http
      * @param IResponse $response Réponse Http
+     * @param int $planningId Contrainte de recherche sur le planning
      *
      * @return IResponse
      * @throws \Exception en cas d'erreur inconnue (fallback, ne doit pas arriver)
      */
-    private function getList(IRequest $request, IResponse $response)
+    private function getList(IRequest $request, IResponse $response, $planningId)
     {
+        $code = -1;
+        $data = [];
         try {
-            $plannings = $this->repository->getList(
-                $request->getQueryParams()
-            );
+            $creneaux = $this->repository->getList(['planningId' => $planningId]);
             $entites = [];
-            foreach ($plannings as $planning) {
-                $entites[] = $this->buildData($planning);
+            foreach ($creneaux as $creneau) {
+                $entites[] = $this->buildData($creneau);
             }
             $code = 200;
             $data = [
@@ -109,7 +119,7 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
     /**
      * Construit le « data » du json
      *
-     * @param Entite $entite Planning
+     * @param Entite $entite Créneau de planning
      *
      * @return array
      */
@@ -117,8 +127,12 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
     {
         return [
             'id' => $entite->getId(),
-            'name' => $entite->getName(),
-            'status' => $entite->getStatus(),
+            'planningId' => $entite->getPlanningId(),
+            'jourId' => $entite->getJourId(),
+            'typeSemaine' => $entite->getTypeSemaine(),
+            'typePeriode' => $entite->getTypePeriode(),
+            'debut' => $entite->getDebut(),
+            'fin' => $entite->getFin(),
         ];
     }
 
@@ -143,15 +157,19 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
         }
 
         try {
-            $planningId = $this->repository->postOne($body, new Entite([]));
+            $creneauxIds = $this->repository->postList($body, new Entite([]));
+            $dataMessage = [];
+            foreach ($creneauxIds as $id) {
+                $dataMessage[] = $this->router->pathFor('getPlanningCreneauDetail', [
+                    'creneauId' => $id,
+                ]);
+            }
             $code = 201;
             $data = [
                 'code' => $code,
                 'status' => 'success',
                 'message' => '',
-                'data' => $this->router->pathFor('getPlanningDetail', [
-                    'planningId' => $planningId
-                ]),
+                'data' => $dataMessage,
             ];
 
             return $response->withJson($data, $code);
@@ -185,17 +203,17 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
             return $this->getResponseBadRequest($response, 'Body request is not a json content');
         }
 
-        $id = (int) $arguments['planningId'];
+        $id = (int) $arguments['creneauId'];
         try {
-            $planning = $this->repository->getOne($id);
+            $creneau = $this->repository->getOne($id);
         } catch (\DomainException $e) {
-            return $this->getResponseNotFound($response, 'Element « plannings#' . $id . ' » is not a valid resource');
+            return $this->getResponseNotFound($response, 'Element « creneaux#' . $id . ' » is not a valid resource');
         } catch (\Exception $e) {
             throw $e;
         }
 
         try {
-            $this->repository->putOne($body, $planning);
+            $this->repository->putOne($body, $creneau);
             $code = 204;
             $data = [
                 'code' => $code,
@@ -209,42 +227,6 @@ final class Controller extends \LibertAPI\Tools\Libraries\AController
             return $this->getResponseMissingArgument($response);
         } catch (\DomainException $e) {
             return $this->getResponseBadDomainArgument($response, $e);
-        } catch (\Exception $e) {
-            throw $e;
-        }
-    }
-
-    /*************************************************
-     * DELETE
-     *************************************************/
-
-     /**
-      * Execute l'ordre HTTP DELETE
-      *
-      * @param IRequest $request Requête Http
-      * @param IResponse $response Réponse Http
-      * @param array $arguments Arguments de route
-      *
-      * @return IResponse
-      * @throws \Exception en cas d'erreur inconnue (fallback, ne doit pas arriver)
-      */
-    public function delete(IRequest $request, IResponse $response, array $arguments)
-    {
-        $id = (int) $arguments['planningId'];
-        try {
-            $planning = $this->repository->getOne($id);
-            $this->repository->deleteOne($planning);
-            $code = 200;
-            $data = [
-                'code' => $code,
-                'status' => 'success',
-                'message' => '',
-                'data' => '',
-            ];
-
-            return $response->withJson($data, $code);
-        } catch (\DomainException $e) {
-            return $this->getResponseNotFound($response, 'Element « plannings#' . $id . ' » is not a valid resource');
         } catch (\Exception $e) {
             throw $e;
         }
