+++ Tools/Route/Authentification.php
@@ -2,4 +2,4 @@
 /*
  * Doit être importé après la création de $app. Ne créé rien.
  */
-$app->get('/authentification', '\LibertAPI\Components\Authentification\Controller:get')->setName('authentification');
+$app->get('/authentification', '\LibertAPI\Authentification\Controller:get')->setName('authentification');
